feat: add cos client

+++ src/main/java/cn/ming/smartmedic/Service/NewsService.java
@@ -13,7 +13,7 @@ public interface NewsService {
 
     List<News> listNews(Map<String, Object> con);
 
-    News getNews(Long id);
+    News getNews(BigInteger id);
 
     Integer saveNews(News news);
